requirements.txt
# pip install -r requirements.txt

opencv-python

#PIL
Pillow

numpy
grid-extractor
